favourites done

+++ src/app/books/book-list/book-item/book-item.component.ts
@@ -2,10 +2,12 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Book } from '../../books.model';
 import { MatDialog } from '@angular/material/dialog';
 import { BookDetailComponent } from '../../book-detail/book-detail.component';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { CartService } from '../../../cart/cart.service';
 import { DataStorageService } from '../../../shared/data-storage.service';
 import { FavouriteService } from '../../../favourites/favourites.service';
+import { AuthService } from '../../../auth/auth.service';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-book-item',
@@ -19,8 +21,11 @@ export class BookItemComponent implements OnInit {
  id: number
  countCartItems: number
  isFavorite: boolean //= false
+ isAuth = false
 
- constructor(public dialog: MatDialog, private route: ActivatedRoute, private cartService: CartService, private dataStore: DataStorageService, private favouriteService: FavouriteService){}
+ userSub: Subscription
+
+ constructor(public dialog: MatDialog, private route: ActivatedRoute, private cartService: CartService, private dataStore: DataStorageService, private favouriteService: FavouriteService, private auth: AuthService,private router: Router){}
  
   ngOnInit(){
     this.route.params.subscribe(params => {
@@ -28,6 +33,10 @@ export class BookItemComponent implements OnInit {
     })
 
     this.isFavorite = this.favouriteService.getFavourites().find(fav => fav.id === this.book.id) ? true : false
+    
+    this.userSub = this.auth.user.subscribe(user => {
+      this.isAuth = !!user;
+    });
   }
 
   openDialog(index: number){
@@ -35,17 +44,22 @@ export class BookItemComponent implements OnInit {
   }
 
   onFavorite(){
-    this.isFavorite = !this.isFavorite
-    
-    if(this.isFavorite){
-      this.isFavorite = true
-      this.favouriteService.addBooksToFavourites(this.book)
-      this.dataStore.storeFavourites()
+    if(this.isAuth){
+      
+      this.isFavorite = !this.isFavorite
+      if(this.isFavorite){
+        this.isFavorite = true
+        this.favouriteService.addBooksToFavourites(this.book)
+        this.dataStore.storeFavourites()
+      }
+      else{
+        this.isFavorite = false
+        this.favouriteService.deleteFavourite(this.id)
+        this.dataStore.storeFavourites()
+      }
     }
     else{
-      this.isFavorite = false
-      this.favouriteService.deleteFavourite(this.id)
-      this.dataStore.storeFavourites()
+      this.router.navigate(['/auth'])
     }
   }
